refactor: 테스트용 Fake 객체들을 repository 및 dispatcher 패키지로 분리
