+++ app/src/test/java/woowacourse/shopping/presentation/cart/viewmodel/CartViewModelTest.kt
@@ -11,7 +11,7 @@ import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.AfterEach
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
-import woowacourse.shopping.fake.FakeCartRepository
+import woowacourse.shopping.fake.repository.FakeCartRepository
 import woowacourse.shopping.fake.fakeProduct
 
 @OptIn(ExperimentalCoroutinesApi::class)
